+++ src/main/java/subway/auth/exception/ExceptionControllerAdvice.java
@@ -1,0 +1,27 @@
+package subway.auth.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import java.sql.SQLException;
+
+@RestControllerAdvice
+public class ExceptionControllerAdvice {
+
+    @ExceptionHandler(InvalidMemberException.class)
+    public ResponseEntity<Void> invalidMember() {
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+    }
+
+    @ExceptionHandler(SQLException.class)
+    public ResponseEntity handleSQLException() {
+        return ResponseEntity.badRequest().build();
+    }
+
+    @ExceptionHandler(InvalidTokenException.class)
+    public ResponseEntity<Void> invalidToken() {
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+    }
+}
